insert registration kind on init as well
This is the only point where kind is set and is used further along in process (eg. submit) to detect what is the kind of the registration in question. Registration cannot carry over from admission to post admission but instead even the same person has to restart the whole registration process in case post admission is opened.