Upgrade dev and test dependencies.
The updated EmbeddedPostgres version ([com.opentable.components/otj-pg-embedded 1.0.0])
uses the testcontainers library to spin up a Dockerized version of Postgres.
This removes the need for custom setup to sniff for open ports AND is also compatible with
the M1 Mac, unlike the older version previously used. As a downside, the Dockerized version
of EmbeddedPostgres likely makes the tests somewhat slower to run.