OPHYKIYP-47: Fix email-queue-reader task not being rescheduled after error
Wrap the whole function body in a try-catch block to ensure that no exception (from eg. the call to pgq/count)
escapes the multimethod implementation.

This is necessary, as the task ultimately gets scheduled using a j.u.c.ScheduledExecutorService instance,
which will stop scheduling a task if an execution of the task throws an exception.