OPHKIOS-67: Speed up exam session listing by adding an index on exam_session_queue.
Running EXPLAIN ANALYZE on the SELECT query used to generate the exam session listing revealed that most of the time
is spent on counting the number of available places in the exam session queue. This seems to be because we lacked
a suitable index on exam_session_queue; instead, the query would use the index on (email, exam_session_id), having
to basically cover the entire table as the column order of the index wasn't suitable.